Different math operant sonar , attempt #2

+++ application/src/main/java/org/togetherjava/tjbot/commands/moderation/modmail/ModMailCommand.java
@@ -103,7 +103,7 @@ public class ModMailCommand extends SlashCommandAdapter {
 
     }
 
-    private long getRemainingTime(MessageChannel messageChannel) {
+    private int getRemainingTime(MessageChannel messageChannel) {
         return COOLDOWN_DURATION_VALUE - getTimePassedSinceCommandCalled(messageChannel);
     }
